Toolbar agenda (#123)
* Show month preview

* Restyle calendar

* Add layouts

* Add sample progress

* Update Kotlin

* Convert ProgressMonthView to Kotlin

* Style progress & indicators

* Change color data

* Colorize progress

* Add initial CreateScheduleFullnessItemsUseCase

* Finish CreateScheduleFullnessItemsUseCase

* Populate month schema

* Limit number of shown tags

* Sort tag colors

* Change schedule summary on swipe

* Add ProgressWeekView

* Set first day of week according to locale

* Optimize xmls

* More xml optimization

* Load schedule for selected date

* Skip first automatic changes in the listeners

* Update toolbar date. Bind event.

* Show day summary

* Add Day items

* Cleanup

* Shift date text up

* Restyle selected date

* Adjust selection box

* WeekBar -> Kotlin

* Add toolbar buttons

* Cleanup

* Store previous date in VS

* Change scroll to today method

* Reset uiVisibility on detach

* Offset end date when setting summary data

* Remove calendarview module

* Cleanup

* Render completed items

* Add swiping

* Bind ViewModels

* Reschedule quest

* Stop listening for changes from DB when Channel is closed

* Do not use ChannelRelay when loading app data

* Cleanup

* Remove ChannelRelay

* Cleanup

* Fix schedule loading

* Listen for schedule summary changes

* Add placeholder Quests to summary

* Fix creating placeholders

* Remove Quest

* Draw border around current date

* Cleanup

* Cleanup

* Extract strings

* Use currentDate in ScheduleVC. Schedule habit updates.

+++ app/src/main/java/io/ipoli/android/common/text/DurationFormatter.kt
@@ -51,7 +51,7 @@ object DurationFormatter {
         }
 
         return if (hours > 0 && mins == 0) {
-            context.resources.getQuantityString(R.plurals.duration_hours, hours.toInt(), hours)
+            context.resources.getQuantityString(R.plurals.duration_hours, hours, hours)
         } else context.resources.getQuantityString(R.plurals.duration_minutes, mins.toInt(), mins)
 
     }
